Add Ticker Scroll Animation
Some of this can be moved to the main creation of the container, but wanted to make sure we knew what the changes were. Creates the max-height of the wrapper, whether animated or not. Makes the scroll containers an inline-block. Sets the animation for the scroll-containers and then sets the animation keyframes. 10px is for the margin-top value, otherwise it jumps at 100%. If that value doesn't have to be hardcoded, that would be best. I don't know how to pull that