Add Ticker Scroll Animation and updateRefreshInterval
New Animation config. Default to off and display the whole wrapper. Set height to maximum that is displayed (starts at value that allows for any size). Overflow will be hidden unless scroll is true. scrollSpeed is seconds per line. 6 = 10 lines shown over 60 seconds. 1 = 10 lines shown over 10 seconds (very fast). In it's current configuration, there can be a slight delay if animation.height is close to the total height of the wrapper. It may take a refresh for it to catch up and start animating. Uses the total height of the wrapper as queried by the document. If you know a better way, please feel free to re-work it. Created a new refresh function to have the refresh match up with the scroll rate. There is a minimum of 2 minutes, but for the animation to not be glitchy, the refresh needs to match when the scrollTop is back to 0. It calculates a multiple of the animationDuration and refreshes at the soonest point after the minimum wait time. e.g 11 lines at 6 seconds is 66 seconds per lap. Update interval is 132 seconds (2x 66sec).